requirements.txt
satcfdi>=4.0.0
python-dotenv>=1.0.1
rich>=13.7.1
pandas>=2.2.0
openpyxl>=3.1.2
python-docx>=1.1.0
streamlit>=1.44.0
plotly>=6.0.0
PyYAML>=6.0.0
lxml>=5.2.2
Pillow>=10.4.0
